Finishing implementation of the hole traveller salesman solution.

+++ src/main/java/br/com/pontoclass/traveller/Route.java
@@ -1,9 +1,11 @@
 package br.com.pontoclass.traveller;
 
+import java.math.BigDecimal;
+
 public interface Route {
 
 	public String getWayString();
 
-	public double getTotalWayCost();
+	public BigDecimal getTotalWayCost();
 
 }
